Database and Monitor for Transactions
Changes to back-end code for establishing connection to database, as well as monitoring unconfirmed transactions. Now, the monitor is able to create get and post requests, as well as authenticate users from being able to access the API endpoint. The data from the largest unconfirmed transactions are stored into the database too.